Simplify R&S
