+++ src/qibolab/instruments/oscillator.py
@@ -39,14 +39,28 @@ class LocalOscillator(Instrument):
     def power(self, x):
         self.settings.power = x
 
+    def upload(self):
+        """Uploads cached setting values to the instruments."""
+
     def connect(self):
         self.is_connected = True
+        self.upload()
 
-    def setup(self, power=None, frequency=None, **kwargs):
-        if power is not None:
-            self.power = power
-        if frequency is not None:
-            self.frequency = frequency
+    def setup(self, **kwargs):
+        """Update instrument settings.
+
+        If the instrument is connected the value is automatically uploaded to the instrument.
+        Otherwise the value is cached and will be uploaded when connection is established.
+
+        Args:
+            **kwargs: Instrument settings loaded from the runcard.
+        """
+        type_ = self.__class__
+        properties = {p for p in dir(type_) if isinstance(getattr(type_, p), property)}
+        for name, value in kwargs.items():
+            if name not in properties:
+                raise KeyError(f"Cannot set {name} to instrument {self.name} of type {type_.__name__}")
+            setattr(self, name, value)
 
     def start(self):
         pass
