+++ src/qibolab/instruments/rohde_schwarz.py
@@ -5,28 +5,19 @@ Supports the following Instruments:
 
 https://qcodes.github.io/Qcodes/api/generated/qcodes.instrument_drivers.rohde_schwarz.html#module-qcodes.instrument_drivers.rohde_schwarz.SGS100A
 """
-from dataclasses import dataclass
-from typing import Optional
-
 import qcodes.instrument_drivers.rohde_schwarz.SGS100A as LO_SGS100A
 from qibo.config import log
 
-from qibolab.instruments.abstract import InstrumentException, InstrumentSettings
-from qibolab.instruments.oscillator import LocalOscillator
-
-
-@dataclass
-class SGS100ASettings(InstrumentSettings):
-    power: Optional[float] = None
-    frequency: Optional[float] = None
-    reference_clock_source: Optional[str] = None
+from qibolab.instruments.abstract import InstrumentException
+from qibolab.instruments.oscillator import LocalOscillator, LocalOscillatorSettings
 
 
 class SGS100A(LocalOscillator):
     def __init__(self, name, address, reference_clock_source="EXT"):
         super().__init__(name, address)
         self.device: LO_SGS100A = None
-        self.settings = SGS100ASettings(reference_clock_source=reference_clock_source)
+        self.settings = LocalOscillatorSettings()
+        self._reference_clock_source = reference_clock_source
 
     @property
     def frequency(self):
@@ -52,14 +43,25 @@ class SGS100A(LocalOscillator):
 
     @property
     def reference_clock_source(self):
-        return self.settings.reference_clock_source
+        return self._reference_clock_source
 
     @reference_clock_source.setter
     def reference_clock_source(self, x):
-        if self.settings.reference_clock_source != x:
-            self.settings.reference_clock_source = x
+        if self.reference_clock_source != x:
+            self._reference_clock_source = x
             if self.is_connected:
                 self.device.ref_osc_source = x
+
+    def upload(self):
+        """Uploads cached setting values to the instruments."""
+        if not self.is_connected:
+            raise RuntimeError("Cannot upload settings if instrument is not connected.")
+
+        if self.settings.frequency is not None:
+            self.device.set("frequency", self.settings.frequency)
+        if self.settings.power is not None:
+            self.device.set("power", self.settings.power)
+        self.device.ref_osc_source = self._reference_clock_source
 
     def connect(self):
         """Connects to the instrument using the IP address set in the runcard."""
@@ -78,24 +80,7 @@ class SGS100A(LocalOscillator):
                 raise InstrumentException(self, f"Unable to connect to {self.name}")
         else:
             raise RuntimeError("There is an open connection to the instrument already")
-        # set instrument frequency and power if they were changed before connecting
-        if self.settings.frequency is not None:
-            self.device.set("frequency", self.settings.frequency)
-        if self.settings.power is not None:
-            self.device.set("power", self.settings.power)
-        self.device.ref_osc_source = self.settings.reference_clock_source
-
-    def setup(self, **kwargs):
-        """Sets instrument parameters.
-
-        If the instrument is connected the value is automatically uploaded to the instrument.
-        Otherwise the value is cached and will be uploaded when connection is established.
-
-        Args:
-            **kwargs: Instrument settings loaded from the runcard.
-        """
-        for name, value in kwargs.items():
-            setattr(self, name, value)
+        self.upload()
 
     def start(self):
         self.device.on()
